Rename UnifinaCoreAPIFilters -> RESTAPIFilters
