+++ grails-app/controllers/com/unifina/controller/ProductApiController.groovy
@@ -1,7 +1,6 @@
 package com.unifina.controller
 
-
-import com.unifina.api.ApiException
+import com.unifina.api.*
 import com.unifina.domain.Permission
 import com.unifina.domain.Product
 import com.unifina.domain.User
