+++ test/unit/com/unifina/controller/StreamApiControllerSpec.groovy
@@ -4,15 +4,15 @@ import com.unifina.ControllerSpecification
 import com.unifina.api.NotFoundException
 import com.unifina.api.NotPermittedException
 import com.unifina.api.ValidationException
+import com.unifina.controller.RESTAPIFilters
 import com.unifina.domain.*
-import com.unifina.filters.UnifinaCoreAPIFilters
 import com.unifina.service.*
 import grails.converters.JSON
 import grails.test.mixin.Mock
 import grails.test.mixin.TestFor
 
 @TestFor(StreamApiController)
-@Mock([User, Stream, Key, Permission, PermissionService, StreamService, DashboardService, IntegrationKey, UnifinaCoreAPIFilters])
+@Mock([User, Stream, Key, Permission, PermissionService, StreamService, DashboardService, IntegrationKey, RESTAPIFilters])
 class StreamApiControllerSpec extends ControllerSpecification {
 
 	User me
